MAT-7855 rename service
